+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 
 # Changelog
 
+### #103 Tech CORE-542 | Change broker commands and sub-commands
+- features:
+    - change clusters cli sub-commands.
+    - create brokers cli command and sub-commands. 
+    - update workspace_init doc to the new broker command.
+- tests:
+    - fixed the tests that are related to brokers and clusters.
+---
+
 ### #104 Tech CORE-545 | external-dns and removal of hostname in cli
 - fix:
     - removed the hostname in the insprctl requests